+++ src/Arrow.jl
@@ -88,6 +88,7 @@ import .ArrowCore: release!
 const AC = ArrowCore
 
 include("ipc_read.jl")
+include("ipc_dictionary.jl")
 include("ipc_write.jl")
 include("cdata.jl")
 include("source.jl")